% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_distance_decay_gravity}
\alias{add_distance_decay_gravity}
\title{Add distance decay to gravity model}
\usage{
add_distance_decay_gravity(gravity_data, distances)
}
\arguments{
\item{gravity_data}{Gravity accessibility data}

\item{distances}{Distance matrix}
}
\value{
Gravity data with distance decay
}
\description{
Add distance decay to gravity model
}
\keyword{internal}
