% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_population_weighting}
\alias{add_population_weighting}
\title{Add population weighting to accessibility results}
\usage{
add_population_weighting(accessibility_results, spatial_zones)
}
\arguments{
\item{accessibility_results}{Accessibility results}

\item{spatial_zones}{Spatial zones with population data}
}
\value{
Accessibility results with population weighting
}
\description{
Add population weighting to accessibility results
}
\keyword{internal}
