% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{add_spatial_features}
\alias{add_spatial_features}
\title{Add spatial features to mobility data}
\usage{
add_spatial_features(mobility_data, spatial_zones)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{Spatial zones}
}
\value{
Data with spatial features added
}
\description{
Add spatial features to mobility data
}
\keyword{internal}
