% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{aggregate_mobility_by_zone}
\alias{aggregate_mobility_by_zone}
\title{Aggregate mobility data by spatial zone}
\usage{
aggregate_mobility_by_zone(mobility_data, spatial_zones, value_column)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{spatial_zones}{SF object with spatial zones}

\item{value_column}{Column to aggregate}
}
\value{
Data frame with aggregated values by zone
}
\description{
Aggregate mobility data by spatial zone
}
\keyword{internal}
