% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{aggregate_temporal_data}
\alias{aggregate_temporal_data}
\title{Aggregate temporal data}
\usage{
aggregate_temporal_data(mobility_data, data_type)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{data_type}{Type of data}
}
\value{
Aggregated data frame
}
\description{
Aggregate temporal data
}
\keyword{internal}
