% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{aggregate_temporal_mobility}
\alias{aggregate_temporal_mobility}
\title{Aggregate mobility data by temporal resolution}
\usage{
aggregate_temporal_mobility(mobility_data, value_column, temporal_resolution)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{value_column}{Column to aggregate}

\item{temporal_resolution}{Time resolution}
}
\value{
Aggregated data frame
}
\description{
Aggregate mobility data by temporal resolution
}
\keyword{internal}
