% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{analyze_mobility_trends}
\alias{analyze_mobility_trends}
\title{Analyze mobility trends}
\usage{
analyze_mobility_trends(ts_object, decomposition)
}
\arguments{
\item{ts_object}{Time series object}

\item{decomposition}{Decomposition results}
}
\value{
Trend analysis results
}
\description{
Analyze mobility trends
}
\keyword{internal}
