% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{analyze_spatial_accessibility}
\alias{analyze_spatial_accessibility}
\title{Analyze spatial accessibility}
\usage{
analyze_spatial_accessibility(spatial_data, spatial_zones)
}
\arguments{
\item{spatial_data}{SF object with aggregated mobility data}

\item{spatial_zones}{SF object with spatial zones}
}
\value{
List with accessibility analysis results
}
\description{
Analyze spatial accessibility
}
\keyword{internal}
