% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{analyze_spatial_autocorrelation}
\alias{analyze_spatial_autocorrelation}
\title{Analyze spatial autocorrelation}
\usage{
analyze_spatial_autocorrelation(spatial_data, weight_matrix)
}
\arguments{
\item{spatial_data}{SF object with aggregated mobility data}

\item{weight_matrix}{Spatial weight matrix}
}
\value{
List with autocorrelation results
}
\description{
Analyze spatial autocorrelation
}
\keyword{internal}
