% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{analyze_spatial_clustering}
\alias{analyze_spatial_clustering}
\title{Analyze spatial clustering of mobility patterns}
\usage{
analyze_spatial_clustering(
  mobility_data,
  spatial_zones,
  clustering_method = c("getis_ord", "moran", "local_moran")
)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{Spatial zones data}

\item{clustering_method}{Method for clustering analysis}
}
\value{
Spatial clustering results
}
\description{
Analyze spatial clustering of mobility patterns
}
\examples{
\dontrun{
# Analyze spatial clustering
clustering <- analyze_spatial_clustering(
  mobility_data = mobility,
  spatial_zones = zones,
  clustering_method = "getis_ord"
)
}
}
