% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{apply_demographic_filters}
\alias{apply_demographic_filters}
\title{Apply demographic filters to mobility data}
\usage{
apply_demographic_filters(mobility_data, demographic_filter)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{demographic_filter}{List of demographic filters}
}
\value{
Filtered data frame
}
\description{
Apply demographic filters to mobility data
}
\keyword{internal}
