% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{apply_distance_decay_weighting}
\alias{apply_distance_decay_weighting}
\title{Apply distance decay weighting}
\usage{
apply_distance_decay_weighting(accessibility_data, distances, value_column)
}
\arguments{
\item{accessibility_data}{Accessibility data}

\item{distances}{Distance matrix}

\item{value_column}{Column to apply decay to}
}
\value{
Data with distance decay applied
}
\description{
Apply distance decay weighting
}
\keyword{internal}
