% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_analysis.R
\name{calculate_containment}
\alias{calculate_containment}
\title{Calculate self-containment index}
\usage{
calculate_containment(od_data)
}
\arguments{
\item{od_data}{Mobility matrix}
}
\value{
Data frame with containment metrics
}
\description{
Calculate self-containment index
}
