% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{calculate_cumulative_accessibility}
\alias{calculate_cumulative_accessibility}
\title{Calculate cumulative accessibility}
\usage{
calculate_cumulative_accessibility(values, distances, threshold)
}
\arguments{
\item{values}{Vector of values (opportunities)}

\item{distances}{Distance matrix}

\item{threshold}{Distance threshold}
}
\value{
Vector of accessibility values
}
\description{
Calculate cumulative accessibility
}
\keyword{internal}
