% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_analysis.R
\name{calculate_distance_decay}
\alias{calculate_distance_decay}
\title{Calculate distance-decay parameters}
\usage{
calculate_distance_decay(od_data, zones, model = "power")
}
\arguments{
\item{od_data}{Mobility matrix}

\item{zones}{Spatial zones with geometry}

\item{model}{Model type: "power" or "exponential"}
}
\value{
List with model parameters and fit statistics
}
\description{
Calculate distance-decay parameters
}
