% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{calculate_gearys_c}
\alias{calculate_gearys_c}
\title{Calculate Geary's C statistic}
\usage{
calculate_gearys_c(values, weight_matrix)
}
\arguments{
\item{values}{Vector of values}

\item{weight_matrix}{Spatial weight matrix}
}
\value{
List with Geary's C statistics
}
\description{
Calculate Geary's C statistic
}
\keyword{internal}
