% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_general_accessibility}
\alias{calculate_general_accessibility}
\title{Calculate general accessibility}
\usage{
calculate_general_accessibility(
  mobility_data,
  spatial_zones,
  distances,
  distance_decay
)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{Spatial zones}

\item{distances}{Distance matrix}

\item{distance_decay}{Apply distance decay}
}
\value{
Accessibility results
}
\description{
Calculate general accessibility
}
\keyword{internal}
