% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_getis_ord_clustering}
\alias{calculate_getis_ord_clustering}
\title{Calculate Getis-Ord clustering statistics}
\usage{
calculate_getis_ord_clustering(zones_with_mobility)
}
\arguments{
\item{zones_with_mobility}{Spatial zones with mobility data}
}
\value{
Getis-Ord clustering results
}
\description{
Calculate Getis-Ord clustering statistics
}
\keyword{internal}
