% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{calculate_gravity_accessibility}
\alias{calculate_gravity_accessibility}
\title{Calculate gravity-based accessibility}
\usage{
calculate_gravity_accessibility(values, distances, decay_parameter = 1.5)
}
\arguments{
\item{values}{Vector of values (opportunities)}

\item{distances}{Distance matrix}

\item{decay_parameter}{Distance decay parameter}
}
\value{
Vector of accessibility values
}
\description{
Calculate gravity-based accessibility
}
\keyword{internal}
