% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_gravity_accessibility_full}
\alias{calculate_gravity_accessibility_full}
\title{Calculate gravity-based accessibility}
\usage{
calculate_gravity_accessibility_full(
  mobility_data,
  spatial_zones,
  distances,
  distance_decay
)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{Spatial zones}

\item{distances}{Distance matrix}

\item{distance_decay}{Apply distance decay}
}
\value{
Gravity accessibility results
}
\description{
Calculate gravity-based accessibility
}
\keyword{internal}
