% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_job_accessibility}
\alias{calculate_job_accessibility}
\title{Calculate job accessibility using commuting patterns}
\usage{
calculate_job_accessibility(
  mobility_data,
  spatial_zones,
  distances,
  distance_decay
)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{Spatial zones}

\item{distances}{Distance matrix}

\item{distance_decay}{Apply distance decay}
}
\value{
Job accessibility results
}
\description{
Calculate job accessibility using commuting patterns
}
\keyword{internal}
