% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{calculate_lisa}
\alias{calculate_lisa}
\title{Calculate Local Indicators of Spatial Association (LISA)}
\usage{
calculate_lisa(values, weight_matrix)
}
\arguments{
\item{values}{Vector of values}

\item{weight_matrix}{Spatial weight matrix}
}
\value{
Vector of local Moran's I values
}
\description{
Calculate Local Indicators of Spatial Association (LISA)
}
\keyword{internal}
