% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{calculate_morans_i}
\alias{calculate_morans_i}
\title{Calculate Moran's I statistic}
\usage{
calculate_morans_i(values, weight_matrix)
}
\arguments{
\item{values}{Vector of values}

\item{weight_matrix}{Spatial weight matrix}
}
\value{
List with Moran's I statistics
}
\description{
Calculate Moran's I statistic
}
\keyword{internal}
