% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{calculate_morans_i_variance}
\alias{calculate_morans_i_variance}
\title{Calculate variance for Moran's I}
\usage{
calculate_morans_i_variance(n, weight_matrix)
}
\arguments{
\item{n}{Number of observations}

\item{weight_matrix}{Spatial weight matrix}
}
\value{
Variance value
}
\description{
Calculate variance for Moran's I
}
\keyword{internal}
