% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{calculate_seasonal_strength}
\alias{calculate_seasonal_strength}
\title{Calculate seasonal strength}
\usage{
calculate_seasonal_strength(decomposition)
}
\arguments{
\item{decomposition}{Decomposition object}
}
\value{
Seasonal strength value
}
\description{
Calculate seasonal strength
}
\keyword{internal}
