% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_zone_distances}
\alias{calculate_zone_distances}
\title{Calculate zone-to-zone distances}
\usage{
calculate_zone_distances(spatial_zones)
}
\arguments{
\item{spatial_zones}{Spatial zones data}
}
\value{
Distance matrix
}
\description{
Calculate zone-to-zone distances
}
\keyword{internal}
