% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_spanish_holidays}
\alias{check_spanish_holidays}
\title{Check for Spanish holidays that may affect mobility patterns}
\usage{
check_spanish_holidays(dates)
}
\arguments{
\item{dates}{Vector of dates to check}
}
\value{
Data frame with holiday information
}
\description{
Check for Spanish holidays that may affect mobility patterns
}
