% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{create_clustered_flow_map}
\alias{create_clustered_flow_map}
\title{Create flow map with clustering for large datasets}
\usage{
create_clustered_flow_map(
  spatial_zones,
  mobility_data,
  cluster_threshold = 1000,
  flow_threshold = 50,
  mapbox_token = NULL,
  title = "Mobility Flows - Clustered"
)
}
\arguments{
\item{spatial_zones}{Spatial zones data (sf object)}

\item{mobility_data}{Mobility data}

\item{cluster_threshold}{Number of flows above which clustering is enabled}

\item{flow_threshold}{Minimum flow threshold to display}

\item{mapbox_token}{Mapbox access token for basemap}

\item{title}{Map title}
}
\value{
Clustered interactive flow map
}
\description{
Create flow map with clustering for large datasets
}
\examples{
\dontrun{
# Create clustered flow map for large datasets
clustered_map <- create_clustered_flow_map(
  spatial_zones = zones,
  mobility_data = mobility,
  cluster_threshold = 1000,
  flow_threshold = 50,
  title = "Madrid Mobility - Clustered View"
)

clustered_map
}
}
