% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_analysis.R
\name{create_district_flow_plot}
\alias{create_district_flow_plot}
\title{Create flow plot showing top destinations/origins}
\usage{
create_district_flow_plot(
  district_flows,
  zones,
  district_id,
  district_name,
  top_n
)
}
\arguments{
\item{district_flows}{Filtered mobility data}

\item{zones}{Spatial zones data}

\item{district_id}{District ID}

\item{district_name}{District name}

\item{top_n}{Number of top flows to show}
}
\value{
ggplot object
}
\description{
Create flow plot showing top destinations/origins
}
