% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{create_district_heatmap}
\alias{create_district_heatmap}
\title{Create heatmap visualization for district mobility}
\usage{
create_district_heatmap(district_flows, district_id, district_name, time_range)
}
\arguments{
\item{district_flows}{Filtered mobility data}

\item{district_id}{District ID}

\item{district_name}{District name}

\item{time_range}{Time range for title}
}
\value{
ggplot object
}
\description{
Create heatmap visualization for district mobility
}
