% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{create_district_timeseries}
\alias{create_district_timeseries}
\title{Create time series plot for district mobility}
\usage{
create_district_timeseries(district_flows, district_id, district_name)
}
\arguments{
\item{district_flows}{Filtered mobility data}

\item{district_id}{District ID}

\item{district_name}{District name}
}
\value{
ggplot object
}
\description{
Create time series plot for district mobility
}
