% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_flow_map}
\alias{create_flow_map}
\title{Create flow map}
\usage{
create_flow_map(od_data, zones, min_flow = 100)
}
\arguments{
\item{od_data}{Aggregated OD data}

\item{zones}{Spatial zones}

\item{min_flow}{Minimum flow to display (default: 100)}
}
\value{
leaflet map
}
\description{
Create flow map
}
