% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{create_mobility_time_series}
\alias{create_mobility_time_series}
\title{Create time series object from mobility data}
\usage{
create_mobility_time_series(ts_data, temporal_resolution)
}
\arguments{
\item{ts_data}{Temporal data frame}

\item{temporal_resolution}{Time resolution}
}
\value{
Time series object
}
\description{
Create time series object from mobility data
}
\keyword{internal}
