% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{create_spatial_weights}
\alias{create_spatial_weights}
\title{Create spatial weight matrix}
\usage{
create_spatial_weights(spatial_zones, distance_threshold)
}
\arguments{
\item{spatial_zones}{SF object with spatial zones}

\item{distance_threshold}{Distance threshold for neighbors}
}
\value{
Spatial weight matrix
}
\description{
Create spatial weight matrix
}
\keyword{internal}
