% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{detect_change_points}
\alias{detect_change_points}
\title{Detect change points in time series}
\usage{
detect_change_points(trend)
}
\arguments{
\item{trend}{Trend values}
}
\value{
List with change point information
}
\description{
Detect change points in time series
}
\keyword{internal}
