% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{detect_data_version}
\alias{detect_data_version}
\title{Detect data version from dates}
\usage{
detect_data_version(dates)
}
\arguments{
\item{dates}{Date vector or range}
}
\value{
Integer version number (1 or 2)
}
\description{
Detect data version from dates
}
