% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{detect_isolation_forest_anomalies}
\alias{detect_isolation_forest_anomalies}
\title{Detect anomalies using Isolation Forest}
\usage{
detect_isolation_forest_anomalies(features, contamination)
}
\arguments{
\item{features}{Feature matrix}

\item{contamination}{Contamination rate}
}
\value{
Anomaly detection results
}
\description{
Detect anomalies using Isolation Forest
}
\keyword{internal}
