% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{detect_spatial_hotspots}
\alias{detect_spatial_hotspots}
\title{Detect spatial hotspots}
\usage{
detect_spatial_hotspots(spatial_data, weight_matrix)
}
\arguments{
\item{spatial_data}{SF object with aggregated mobility data}

\item{weight_matrix}{Spatial weight matrix}
}
\value{
List with hotspot detection results
}
\description{
Detect spatial hotspots
}
\keyword{internal}
