% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{detect_statistical_anomalies}
\alias{detect_statistical_anomalies}
\title{Detect anomalies using statistical methods}
\usage{
detect_statistical_anomalies(features, contamination)
}
\arguments{
\item{features}{Feature matrix}

\item{contamination}{Contamination rate}
}
\value{
Anomaly detection results
}
\description{
Detect anomalies using statistical methods
}
\keyword{internal}
