% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{detect_svm_anomalies}
\alias{detect_svm_anomalies}
\title{Detect anomalies using One-Class SVM}
\usage{
detect_svm_anomalies(features, contamination)
}
\arguments{
\item{features}{Feature matrix}

\item{contamination}{Contamination rate}
}
\value{
Anomaly detection results
}
\description{
Detect anomalies using One-Class SVM
}
\keyword{internal}
