% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{detect_time_series_anomalies}
\alias{detect_time_series_anomalies}
\title{Detect time series anomalies}
\usage{
detect_time_series_anomalies(ts_object, decomposition)
}
\arguments{
\item{ts_object}{Time series object}

\item{decomposition}{Decomposition results}
}
\value{
Anomaly detection results
}
\description{
Detect time series anomalies
}
\keyword{internal}
