% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{estimate_data_size}
\alias{estimate_data_size}
\title{Estimate data size for planning}
\usage{
estimate_data_size(n_days, version)
}
\arguments{
\item{n_days}{Number of days}

\item{version}{Data version}
}
\value{
List with size estimates
}
\description{
Estimate data size for planning
}
\keyword{internal}
