% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{export_flow_map}
\alias{export_flow_map}
\title{Export interactive flow map to HTML file}
\usage{
export_flow_map(
  flow_map,
  filename = "mobility_flows.html",
  selfcontained = TRUE
)
}
\arguments{
\item{flow_map}{Flow map object created by create_interactive_flow_map}

\item{filename}{Output HTML filename}

\item{selfcontained}{Whether to create self-contained HTML file}
}
\value{
Path to saved HTML file
}
\description{
Export interactive flow map to HTML file
}
\examples{
\dontrun{
# Create and export flow map
flow_map <- create_interactive_flow_map(zones, mobility)
export_flow_map(flow_map, "madrid_flows.html")
}
}
