% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{generate_validation_recommendations}
\alias{generate_validation_recommendations}
\title{Generate validation recommendations}
\usage{
generate_validation_recommendations(validation_results, version)
}
\arguments{
\item{validation_results}{List with all validation results}

\item{version}{Data version}
}
\value{
Character vector with recommendations
}
\description{
Generate validation recommendations
}
\keyword{internal}
