% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_analysis_recommendations}
\alias{get_analysis_recommendations}
\title{Get recommended analysis approaches for data characteristics}
\usage{
get_analysis_recommendations(data_characteristics)
}
\arguments{
\item{data_characteristics}{List with data properties}
}
\value{
List with recommended analysis methods
}
\description{
Get recommended analysis approaches for data characteristics
}
