% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_key}
\alias{get_cache_key}
\title{Cache management utilities}
\usage{
get_cache_key(dates, level, time_window = NULL)
}
\description{
Cache management utilities
}
\keyword{Get}
\keyword{cache}
\keyword{data}
\keyword{for}
\keyword{internal}
\keyword{key}
\keyword{request}
