% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_current_data_version}
\alias{get_current_data_version}
\title{Get current data version}
\usage{
get_current_data_version()
}
\value{
Current data version (1 or 2)
}
\description{
Get current data version
}
