% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{get_data_from_database}
\alias{get_data_from_database}
\title{Get data from DuckDB database}
\usage{
get_data_from_database(data_type, zone_level, dates, version)
}
\arguments{
\item{data_type}{Type of data}

\item{zone_level}{Spatial level}

\item{dates}{Processed dates}

\item{version}{Data version}
}
\value{
Data frame with mobility data
}
\description{
Get data from DuckDB database
}
\keyword{internal}
