% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_type_metadata}
\alias{get_data_type_metadata}
\title{Get metadata for specific data type and version}
\usage{
get_data_type_metadata(data_type = "od", version = 2, zone_level = "districts")
}
\arguments{
\item{data_type}{Type of data ("od", "nt", "os")}

\item{version}{Data version (1 or 2)}

\item{zone_level}{Spatial level ("districts", "municipalities", "luas")}
}
\value{
List with detailed metadata
}
\description{
Get metadata for specific data type and version
}
