% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_version_info}
\alias{get_data_version_info}
\title{Get comprehensive data version information}
\usage{
get_data_version_info(version = NULL, dates = NULL)
}
\arguments{
\item{version}{Data version (1 or 2). If NULL, detects from dates}

\item{dates}{Date range to help determine version}
}
\value{
List with detailed metadata about data version capabilities
}
\description{
Get comprehensive data version information
}
\examples{
\dontrun{
# Get v1 data info
v1_info <- get_data_version_info(version = 1)
print(v1_info$description)
print(v1_info$available_variables)

# Get v2 data info
v2_info <- get_data_version_info(version = 2)
print(v2_info$demographic_variables)
print(v2_info$use_cases)
}
}
