% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{get_data_with_fallback}
\alias{get_data_with_fallback}
\title{Get data with intelligent fallback strategies}
\usage{
get_data_with_fallback(data_type, zone_level, dates, version, use_cache)
}
\arguments{
\item{data_type}{Type of data}

\item{zone_level}{Spatial level}

\item{dates}{Processed dates}

\item{version}{Data version}

\item{use_cache}{Whether to use cache}
}
\value{
Data frame with mobility data
}
\description{
Get data with intelligent fallback strategies
}
\keyword{internal}
