% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_analysis.R
\name{get_mobility_matrix}
\alias{get_mobility_matrix}
\title{Get mobility matrix}
\usage{
get_mobility_matrix(dates, level = "dist", time_window = NULL)
}
\arguments{
\item{dates}{Date range (character vector of length 2)}

\item{level}{Spatial level ("dist", "muni", "lua")}

\item{time_window}{Hour range (e.g. c(7,9))}
}
\value{
Data frame with mobility data
}
\description{
Get mobility matrix
}
