% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_optimal_parameters}
\alias{get_optimal_parameters}
\title{Get optimal analysis parameters based on MITMA data characteristics}
\usage{
get_optimal_parameters(analysis_type = "exploratory", data_size = "medium")
}
\arguments{
\item{analysis_type}{Type of analysis: "exploratory", "detailed", "regional", "temporal"}

\item{data_size}{Expected data size: "small", "medium", "large"}
}
\value{
List with recommended parameters
}
\description{
Get optimal analysis parameters based on MITMA data characteristics
}
