% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{get_processing_recommendations}
\alias{get_processing_recommendations}
\title{Get processing recommendations}
\usage{
get_processing_recommendations(n_days, version)
}
\arguments{
\item{n_days}{Number of days}

\item{version}{Data version}
}
\value{
List with processing recommendations
}
\description{
Get processing recommendations
}
\keyword{internal}
