% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_hierarchy.R
\name{get_spatial_zones}
\alias{get_spatial_zones}
\title{Get spatial zones with parsed metadata}
\usage{
get_spatial_zones(level = "dist", version = 2)
}
\arguments{
\item{level}{One of: "dist", "muni", "lua"}

\item{version}{Data version (default: 2)}
}
\value{
sf object
}
\description{
Get spatial zones with parsed metadata
}
