% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{interpret_spatial_autocorrelation}
\alias{interpret_spatial_autocorrelation}
\title{Interpret spatial autocorrelation results}
\usage{
interpret_spatial_autocorrelation(moran_i, geary_c)
}
\arguments{
\item{moran_i}{Moran's I results}

\item{geary_c}{Geary's C results}
}
\value{
Character description of spatial pattern
}
\description{
Interpret spatial autocorrelation results
}
\keyword{internal}
