% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{make_model_predictions}
\alias{make_model_predictions}
\title{Make predictions with trained model}
\usage{
make_model_predictions(model, X_predict, model_type)
}
\arguments{
\item{model}{Trained model object}

\item{X_predict}{Features for prediction}

\item{model_type}{Type of model}
}
\value{
Predictions with confidence intervals
}
\description{
Make predictions with trained model
}
\keyword{internal}
