% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mobspain_status}
\alias{mobspain_status}
\title{Get package status and diagnostics}
\usage{
mobspain_status()
}
\value{
List with package status information
}
\description{
Get package status and diagnostics
}
