% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{perform_seasonal_decomposition}
\alias{perform_seasonal_decomposition}
\title{Perform seasonal decomposition}
\usage{
perform_seasonal_decomposition(ts_object, method)
}
\arguments{
\item{ts_object}{Time series object}

\item{method}{Decomposition method}
}
\value{
Decomposition results
}
\description{
Perform seasonal decomposition
}
\keyword{internal}
