% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot.mobspain_time_series}
\alias{plot.mobspain_time_series}
\title{Plot method for mobspain time series analysis}
\usage{
\method{plot}{mobspain_time_series}(x, ...)
}
\arguments{
\item{x}{mobspain_time_series object}

\item{...}{Additional plotting parameters}
}
\value{
ggplot object
}
\description{
Plot method for mobspain time series analysis
}
