% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_distance_decay}
\alias{plot_distance_decay}
\title{Plot distance decay relationship}
\usage{
plot_distance_decay(decay_result, log_scale = TRUE)
}
\arguments{
\item{decay_result}{Result from calculate_distance_decay()}

\item{log_scale}{Use log scale for axes (default: TRUE)}
}
\value{
ggplot object
}
\description{
Plot distance decay relationship
}
