% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{predict_mobility_patterns}
\alias{predict_mobility_patterns}
\title{Machine Learning and Predictive Analytics for Spanish Mobility Data}
\usage{
predict_mobility_patterns(
  mobility_data,
  prediction_dates,
  features = c("hour", "weekday", "distance"),
  model_type = c("random_forest", "linear_regression", "xgboost"),
  spatial_zones = NULL
)
}
\arguments{
\item{mobility_data}{Historical mobility data for training}

\item{prediction_dates}{Dates to predict mobility for}

\item{features}{Features to use for prediction}

\item{model_type}{Type of model to use}

\item{spatial_zones}{Spatial zones data for spatial features}
}
\value{
Mobility predictions with confidence intervals
}
\description{
This file contains machine learning functions for mobility prediction,
pattern recognition, and advanced analytics.
Predict mobility patterns using machine learning
}
\examples{
\dontrun{
# Predict mobility patterns
predictions <- predict_mobility_patterns(
  mobility_data = historical_mobility,
  prediction_dates = c("2024-01-15", "2024-01-16"),
  model_type = "random_forest",
  features = c("hour", "weekday", "distance", "activity_origin")
)
}
}
