% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{prepare_anomaly_features}
\alias{prepare_anomaly_features}
\title{Prepare features for anomaly detection}
\usage{
prepare_anomaly_features(mobility_data)
}
\arguments{
\item{mobility_data}{Mobility data}
}
\value{
Feature matrix for anomaly detection
}
\description{
Prepare features for anomaly detection
}
\keyword{internal}
