% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{prepare_filter_options}
\alias{prepare_filter_options}
\title{Prepare filter options for dashboard}
\usage{
prepare_filter_options(mobility_data)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}
}
\value{
List with filter options
}
\description{
Prepare filter options for dashboard
}
\keyword{internal}
