% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{prepare_flow_map_flows}
\alias{prepare_flow_map_flows}
\title{Prepare flows data for flow map}
\usage{
prepare_flow_map_flows(mobility_data, flow_threshold, time_enabled)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{flow_threshold}{Minimum flow threshold}

\item{time_enabled}{Include time information}
}
\value{
Data frame with flow information
}
\description{
Prepare flows data for flow map
}
\keyword{internal}
