% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{prepare_flow_map_locations}
\alias{prepare_flow_map_locations}
\title{Prepare locations data for flow map}
\usage{
prepare_flow_map_locations(spatial_zones)
}
\arguments{
\item{spatial_zones}{Spatial zones data}
}
\value{
Data frame with location information
}
\description{
Prepare locations data for flow map
}
\keyword{internal}
