% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{prepare_ml_prediction_data}
\alias{prepare_ml_prediction_data}
\title{Prepare prediction data}
\usage{
prepare_ml_prediction_data(
  mobility_data,
  prediction_dates,
  features,
  spatial_zones
)
}
\arguments{
\item{mobility_data}{Historical mobility data}

\item{prediction_dates}{Dates to predict}

\item{features}{Features to use}

\item{spatial_zones}{Spatial zones}
}
\value{
Prepared prediction data
}
\description{
Prepare prediction data
}
\keyword{internal}
