% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{prepare_ml_training_data}
\alias{prepare_ml_training_data}
\title{Prepare training data for machine learning}
\usage{
prepare_ml_training_data(mobility_data, features, spatial_zones)
}
\arguments{
\item{mobility_data}{Raw mobility data}

\item{features}{Features to extract}

\item{spatial_zones}{Spatial zones for spatial features}
}
\value{
Prepared training data
}
\description{
Prepare training data for machine learning
}
\keyword{internal}
