% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{prepare_spatial_data}
\alias{prepare_spatial_data}
\title{Prepare spatial data for dashboard}
\usage{
prepare_spatial_data(mobility_data, spatial_zones)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{spatial_zones}{SF object with spatial zones}
}
\value{
SF object with aggregated spatial data
}
\description{
Prepare spatial data for dashboard
}
\keyword{internal}
