% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{prepare_temporal_data}
\alias{prepare_temporal_data}
\title{Prepare temporal data for dashboard}
\usage{
prepare_temporal_data(mobility_data)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}
}
\value{
Data frame with temporal aggregations
}
\description{
Prepare temporal data for dashboard
}
\keyword{internal}
