% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{print.district_analysis}
\alias{print.district_analysis}
\title{Print method for district analysis}
\usage{
\method{print}{district_analysis}(x, ...)
}
\arguments{
\item{x}{district_analysis object}

\item{...}{Additional arguments}
}
\description{
Print method for district analysis
}
