% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{print.mobility_anomalies}
\alias{print.mobility_anomalies}
\title{Print method for mobility anomalies}
\usage{
\method{print}{mobility_anomalies}(x, ...)
}
\arguments{
\item{x}{Mobility anomalies object}

\item{...}{Additional arguments}
}
\description{
Print method for mobility anomalies
}
