% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{print.mobility_predictions}
\alias{print.mobility_predictions}
\title{Print method for mobility predictions}
\usage{
\method{print}{mobility_predictions}(x, ...)
}
\arguments{
\item{x}{Mobility predictions object}

\item{...}{Additional arguments}
}
\description{
Print method for mobility predictions
}
