% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print.mobility_viz_suite}
\alias{print.mobility_viz_suite}
\title{Print method for mobility visualization suite}
\usage{
\method{print}{mobility_viz_suite}(x, ...)
}
\arguments{
\item{x}{mobility_viz_suite object}

\item{...}{Additional arguments}
}
\description{
Print method for mobility visualization suite
}
