% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.mobspain_status}
\alias{print.mobspain_status}
\title{Print method for mobspain_status}
\usage{
\method{print}{mobspain_status}(x, ...)
}
\arguments{
\item{x}{mobspain_status object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for mobspain_status
}
