% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{print.mobspain_time_series}
\alias{print.mobspain_time_series}
\title{Print method for mobspain time series analysis}
\usage{
\method{print}{mobspain_time_series}(x, ...)
}
\arguments{
\item{x}{mobspain_time_series object}

\item{...}{Additional parameters}
}
\description{
Print method for mobspain time series analysis
}
