% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.spanish_mobility_validation}
\alias{print.spanish_mobility_validation}
\title{Print method for Spanish mobility validation results}
\usage{
\method{print}{spanish_mobility_validation}(x, ...)
}
\arguments{
\item{x}{Spanish mobility validation object}

\item{...}{Additional arguments}
}
\description{
Print method for Spanish mobility validation results
}
