% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{process_dates_for_spanishoddata}
\alias{process_dates_for_spanishoddata}
\title{Process dates for spanishoddata compatibility}
\usage{
process_dates_for_spanishoddata(dates)
}
\arguments{
\item{dates}{Date input in various formats}
}
\value{
Processed dates for spanishoddata
}
\description{
Process dates for spanishoddata compatibility
}
\keyword{internal}
