% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{save_dashboard}
\alias{save_dashboard}
\title{Save dashboard to HTML file}
\usage{
save_dashboard(dashboard_components, save_path)
}
\arguments{
\item{dashboard_components}{Dashboard components}

\item{save_path}{Path to save file}
}
\description{
Save dashboard to HTML file
}
\keyword{internal}
