% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save data to cache}
\usage{
save_to_cache(data, cache_key)
}
\arguments{
\item{data}{Data to cache}

\item{cache_key}{Character string identifying the cache file}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Save data to cache
}
