% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{train_linear_regression_model}
\alias{train_linear_regression_model}
\title{Train linear regression model}
\usage{
train_linear_regression_model(X_train, y_train)
}
\arguments{
\item{X_train}{Training features}

\item{y_train}{Training target}
}
\value{
Trained linear regression model
}
\description{
Train linear regression model
}
\keyword{internal}
