% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{train_random_forest_model}
\alias{train_random_forest_model}
\title{Train random forest model}
\usage{
train_random_forest_model(X_train, y_train)
}
\arguments{
\item{X_train}{Training features}

\item{y_train}{Training target}
}
\value{
Trained random forest model
}
\description{
Train random forest model
}
\keyword{internal}
