% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{train_xgboost_model}
\alias{train_xgboost_model}
\title{Train XGBoost model}
\usage{
train_xgboost_model(X_train, y_train)
}
\arguments{
\item{X_train}{Training features}

\item{y_train}{Training target}
}
\value{
Trained XGBoost model
}
\description{
Train XGBoost model
}
\keyword{internal}
