% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data_anomalies}
\alias{validate_data_anomalies}
\title{Validate data for statistical anomalies}
\usage{
validate_data_anomalies(mobility_data, version)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{version}{Data version}
}
\value{
List with anomaly validation results
}
\description{
Validate data for statistical anomalies
}
\keyword{internal}
