% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data_completeness}
\alias{validate_data_completeness}
\title{Validate data completeness}
\usage{
validate_data_completeness(mobility_data, version)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{version}{Data version}
}
\value{
List with completeness validation results
}
\description{
Validate data completeness
}
\keyword{internal}
