% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Validate data structure against Spanish mobility data specifications}
\usage{
validate_data_structure(mobility_data, version)
}
\arguments{
\item{mobility_data}{Data frame with mobility data}

\item{version}{Data version (1 or 2)}
}
\value{
List with structure validation results
}
\description{
Validate data structure against Spanish mobility data specifications
}
\keyword{internal}
