% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_dates}
\alias{validate_dates}
\title{Enhanced validation utilities}
\usage{
validate_dates(dates)
}
\description{
Enhanced validation utilities
}
\keyword{Validate}
\keyword{date}
\keyword{inputs}
\keyword{internal}
