% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{validate_enhanced_data_request}
\alias{validate_enhanced_data_request}
\title{Validate enhanced data request}
\usage{
validate_enhanced_data_request(
  data_type,
  zone_level,
  version,
  demographic_filter
)
}
\arguments{
\item{data_type}{Type of data}

\item{zone_level}{Spatial level}

\item{version}{Data version}

\item{demographic_filter}{Demographic filters}
}
\description{
Validate enhanced data request
}
\keyword{internal}
