% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_time_window}
\alias{validate_time_window}
\title{Validate time window}
\usage{
validate_time_window(time_window)
}
\arguments{
\item{time_window}{Time window to validate}
}
\value{
Validated time window
}
\description{
Validate time window
}
\keyword{internal}
