% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{connect_mobility_db}
\alias{connect_mobility_db}
\title{Connect to mobility database}
\usage{
connect_mobility_db()
}
\value{
DuckDB connection object
}
\description{
Establishes connection to DuckDB database with processed mobility data
}
