% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_viz.R
\name{export_visualizations}
\alias{export_visualizations}
\title{Export visualizations to files}
\usage{
export_visualizations(
  viz_suite,
  output_dir = ".",
  formats = c("png", "html"),
  width = 10,
  height = 8,
  dpi = 300
)
}
\arguments{
\item{viz_suite}{mobility_viz_suite object}

\item{output_dir}{Output directory}

\item{formats}{File formats to export: "png", "pdf", "html"}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}

\item{dpi}{Resolution for raster formats}
}
\description{
Export visualizations to files
}
