% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_data_version_info}
\alias{get_data_version_info}
\title{Get information about Spanish mobility data versions}
\usage{
get_data_version_info()
}
\value{
List with detailed information about both data versions
}
\description{
Get information about Spanish mobility data versions
}
\examples{
# Get comprehensive version information
version_info <- get_data_version_info()
version_info$comparison
}
