% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_daily_mobility}
\alias{plot_daily_mobility}
\title{Plot daily mobility patterns}
\usage{
plot_daily_mobility(od_data)
}
\arguments{
\item{od_data}{Mobility matrix}
}
\value{
ggplot object
}
\description{
Plot daily mobility patterns
}
