% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mobility_heatmap}
\alias{plot_mobility_heatmap}
\title{Plot mobility heatmap}
\usage{
plot_mobility_heatmap(od_data, top_n = 50)
}
\arguments{
\item{od_data}{Mobility matrix}

\item{top_n}{Number of top flows to show (default: 50)}
}
\value{
ggplot object
}
\description{
Plot mobility heatmap
}
