% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_dates}
\alias{validate_dates}
\title{Validate date inputs}
\usage{
validate_dates(dates)
}
\arguments{
\item{dates}{Date vector to validate}
}
\value{
Validated date vector
}
\description{
Validate date inputs
}
\keyword{internal}
